Integrate installed tool manager view into toolshed component
