+++ client/galaxy/scripts/components/Toolshed/RepositoryDetails/InstallationButton.vue
