+++ client/galaxy/scripts/components/Toolshed/RepositoryDetails/RepositoryTools.vue
