+++ client/galaxy/scripts/components/Toolshed/SearchList/Index.vue
@@ -1,0 +1,90 @@
+<template>
+    <div>
+        <serverselection
+            :toolshedUrl="toolshedUrl"
+            :toolshedUrls="toolshedUrls"
+            :total="total"
+            :loading="loading"
+            @onToolshed="setToolshed"
+        />
+        <repositories
+            :query="query"
+            :scrolled="scrolled"
+            :toolshedUrl="toolshedUrl"
+            @onError="setError"
+            v-if="!queryEmpty"
+        />
+        <categories
+            :toolshedUrl="toolshedUrl"
+            :loading="loading"
+            @onCategory="setQuery"
+            @onTotal="setTotal"
+            @onError="setError"
+            @onLoading="setLoading"
+            v-show="queryEmpty"
+        />
+    </div>
+</template>
+<script>
+import { getGalaxyInstance } from "app";
+import Categories from "./Categories.vue";
+import Repositories from "./Repositories.vue";
+import ServerSelection from "./ServerSelection.vue";
+export default {
+    props: ["query", "scrolled"],
+    components: {
+        categories: Categories,
+        repositories: Repositories,
+        serverselection: ServerSelection
+    },
+    data() {
+        return {
+            toolshedUrl: null,
+            toolshedUrls: [],
+            queryLength: 3,
+            loading: false,
+            total: 0,
+            error: null,
+            tabCurrent: 'true',
+            tabOptions: [
+              { text: 'Search All', value: true },
+              { text: 'Installed Only', value: false }
+            ]
+        };
+    },
+    created() {
+        this.configureToolsheds();
+    },
+    computed: {
+        queryEmpty() {
+            return !this.query || this.query.length < this.queryLength;
+        }
+    },
+    methods: {
+        configureToolsheds() {
+            const galaxy = getGalaxyInstance();
+            this.toolshedUrls = galaxy.config.tool_shed_urls;
+            if (!this.toolshedUrls || this.toolshedUrls.length == 0) {
+                this.setError("Toolshed registry is empty, no servers found.");
+            } else {
+                this.toolshedUrl = this.toolshedUrls[0];
+            }
+        },
+        setError(error) {
+            this.$emit("onError", error);
+        },
+        setQuery(query) {
+            this.$emit("onQuery", query);
+        },
+        setToolshed(url) {
+            this.toolshedUrl = url;
+        },
+        setTotal(total) {
+            this.total = total;
+        },
+        setLoading(loading) {
+            this.loading = loading;
+        }
+    }
+};
+</script>
